+++ priv/static/phoenix_live_view.esm.js
@@ -3454,6 +3454,11 @@ var LiveSocket = class {
       return;
     }
     this.boundTopLevelEvents = true;
+    this.socket.onClose((event) => {
+      if (event.code === 1e3 && this.main) {
+        this.reloadWithJitter(this.main);
+      }
+    });
     document.body.addEventListener("click", function() {
     });
     window.addEventListener("pageshow", (e) => {