+++ priv/static/phoenix_live_view.js
@@ -3484,6 +3484,11 @@ within:
         return;
       }
       this.boundTopLevelEvents = true;
+      this.socket.onClose((event) => {
+        if (event.code === 1e3 && this.main) {
+          this.reloadWithJitter(this.main);
+        }
+      });
       document.body.addEventListener("click", function() {
       });
       window.addEventListener("pageshow", (e) => {